Update website meta for social graph meta tags.
In this pull request, a description used for social sharing of this
website is added in `config/_default/params.toml`.

The former solution did not work because variables defined in
`config/_default` always have the highest priority, which means they
will overwrite any variable defined in `config.toml` that has the same
name.